feat(api): sendgrid_mail update + removed sendgrid_mail_mailing_list
