+++ api/.env-dist
@@ -32,11 +32,6 @@ SENDGRID_SENDER=YOUR_SENGRID_SENDER
 SENDGRID_MFA_TEMPLATE_ID=YOUR_SENGRID_MFA_TEMPLATE_ID
 HUNTER_API_KEY=YOUR_HUNTER_KEY
 
-# email
-DEFAULT_MAIL_SENDER=your_default_mail_sender
-SENDGRID_MAIL_API_KEY=your_sendgrid_mail_api_key
-TEMPLATE_MAIL_ID=your_template_mail_id
-
 # API linter stuff (you don't actually need this to run the app, just some for linter)
 DJANGO_SETTINGS_MODULE='core.settings'
